feat(messages): fold /clinic/morning-brief into Smart Inbox as filter=brief (EMR-708)
Merges the Morning Brief view into /clinic/messages under a brief
filter category. Existing bookmarks to /clinic/morning-brief redirect.

- Extracts the brief content into messages/brief-view.tsx (no semantic
  change to data fetching or render).
- /clinic/morning-brief/page.tsx is now a server-side redirect to
  /clinic/messages?filter=brief.
- messages/page.tsx accepts ?filter=brief and renders the brief view
  with a filter-chip header; default (no filter) shows the Smart Inbox
  with the same chip group so users can flip into Brief.
